pre-view dash

+++ src/app/services/agent/agent.service.ts
@@ -4,6 +4,7 @@ import {Cookie} from 'ng2-cookies';
 import {Agent} from '../../models/Agent';
 import {FormArray} from '@angular/forms';
 import {Observable} from 'rxjs/Observable';
+import {Error} from "../../models/Error";
 
 @Injectable()
 export class AgentService {
@@ -14,29 +15,31 @@ export class AgentService {
     const url = 'http://localhost:8081/agents/';
     const headers = new HttpHeaders({Authorization: Cookie.get('token'), 'Content-Type': 'application/json'});
     return new Promise((resolve, reject) => {
-      this.http.get(url, {headers: headers}).toPromise()
-        .then(response => {
-          resolve(response);
-        })
-        .catch(error => {
-          reject(error);
-        });
+      if(!Cookie.get('token')) {
+        return reject(new Error('Unauthorized user','401 Unauthorized', 401));
+      }
+      else {
+        this.http.get(url, {headers: headers}).toPromise()
+          .then(response => {
+            resolve(response);
+          })
+          .catch(error => {
+            reject(error);
+          });
+      }
     });
   }
-
-
-  /*getAllAgents(): Observable<any> {
-    const url = 'http://localhost:8081/agents/';
-    const headers = new HttpHeaders({Authorization: Cookie.get('token'), 'Content-Type': 'application/json'});
-
-     return this.http.get(url, {headers: headers});
-
-  }*/
 
   getAgentById(id: number): Promise<any> {
     const url = 'http://localhost:8081/agents/' + id;
     const headers = new HttpHeaders({Authorization: Cookie.get('token'), 'Content-Type': 'application/json'});
     return new Promise((resolve, reject) => {
+      if(!Cookie.get('token')) {
+        return reject(new Error('Unauthorized user','401 Unauthorized', 401));
+      }
+      if(id < 0) {
+        return reject(new Error('Unprocessable Entity','422 Unprocessable Entity', 422));
+      }
       this.http.get(url, {headers: headers}).toPromise()
         .then(response => {
           resolve(response);
@@ -51,6 +54,12 @@ export class AgentService {
     const url = 'http://localhost:8081/agents/';
     const headers = new HttpHeaders({Authorization: Cookie.get('token'), 'Content-Type': 'application/json'});
     return new Promise((resolve, reject) => {
+      if(!Cookie.get('token')) {
+        return reject(new Error('Unauthorized user','401 Unauthorized', 401));
+      }
+      if(!agent) {
+        return reject(new Error('Conflict','409 Conflict', 409));
+      }
       this.http.post(url, agent, {headers: headers}).toPromise()
         .then(response => {
           resolve(response);
@@ -65,6 +74,15 @@ export class AgentService {
     const url = 'http://localhost:8081/agents/' + id;
     const headers = new HttpHeaders({Authorization: Cookie.get('token'), 'Content-Type': 'application/json'});
     return new Promise((resolve, reject) => {
+      if(!Cookie.get('token')) {
+        return reject(new Error('Unauthorized user','401 Unauthorized', 401));
+      }
+      if(id < 0) {
+        return reject(new Error('Unprocessable Entity','422 Unprocessable Entity', 422));
+      }
+      if(!agent) {
+        return reject(new Error('Conflict','409 Conflict', 409));
+      }
       this.http.put(url, agent, {headers: headers}).toPromise()
         .then(response => {
           resolve(response);
@@ -95,6 +113,12 @@ export class AgentService {
     const url = 'http://localhost:8081/agents/' + id;
     const headers = new HttpHeaders({Authorization: Cookie.get('token'), 'Content-Type': 'application/json'});
     return new Promise((resolve, reject) => {
+      if(!Cookie.get('token')) {
+        return reject(new Error('Unauthorized user','401 Unauthorized', 401));
+      }
+      if(id < 0) {
+        return reject(new Error('Unprocessable Entity','422 Unprocessable Entity', 422));
+      }
       this.http.delete(url, {headers: headers}).toPromise()
         .then(response => {
           resolve(response);
@@ -110,6 +134,7 @@ export class AgentService {
     for (let i = 0; i < array.length; i++) {
       promises.push(this.deleteAgent(array[i].id));
     }
+
     return new Promise((resolve, reject) => {
       Promise.all(promises)
         .then(data => {
